requirements.txt
streamlit>=1.30.0
openai>=1.2.4
google-generativeai>=0.3.1
python-docx>=1.0.0
requests>=2.31.0

